+++ flutter-hms-health/android/src/main/java/com/huawei/hms/flutter/health/modules/datacontroller/service/DataControllerService.java
@@ -26,6 +26,8 @@ import com.huawei.hms.hihealth.options.ReadOptions;
 import com.huawei.hms.hihealth.options.UpdateOptions;
 import com.huawei.hms.hihealth.result.ReadReply;
 
+import java.util.List;
+
 public interface DataControllerService {
     /**
      * Insert the user's fitness and health data into the Health platform.
@@ -94,6 +96,16 @@ public interface DataControllerService {
         final ResultListener<SampleSet> dataResultListener);
 
     /**
+     * Querying the Summary Fitness and Health Data of the User of the Current day for multiple data types.
+     *
+     * @param dataController {@link DataController} instance.
+     * @param dataTypes {@link List<DataType>} instance.
+     * @param dataResultListener {@link VoidResultListener} listener.
+     */
+    void readTodayList(final DataController dataController, final List<DataType> dataTypes,
+        final ResultListener<List> dataResultListener);
+
+    /**
      * Querying the Summary Fitness and Health Data of the User between selected dates.
      *
      * @param dataController {@link DataController} instance.
@@ -104,6 +116,18 @@ public interface DataControllerService {
      */
     void readDailySummation(DataController dataController, DataType dataType, int startTime, int endTime,
         final ResultListener<SampleSet> dataResultListener);
+
+    /**
+     * Querying the Summary Fitness and Health Data of the User between selected dates.
+     *
+     * @param dataController {@link DataController} instance.
+     * @param dataTypes {@link List<DataType>} instance.
+     * @param startTime An 8-digit integer in the format of YYYYMMDD, for example, 20200803.
+     * @param endTime An 8-digit integer in the format of YYYYMMDD, for example, 20200903.
+     * @param dataResultListener {@link ResultListener } listener.
+     */
+    void readDailySummationList(DataController dataController, List<DataType> dataTypes, int startTime, int endTime,
+        final ResultListener<List> dataResultListener);
 
     /**
      * Clearing the User's Fitness and Health Data from the Device and Cloud