+++ flutter-hms-health/lib/src/options/hihealth_option.dart
@@ -16,6 +16,7 @@
 
 part of huawei_health;
 
+@Deprecated('This Class has been deprecated.')
 /// Defines the request permissions.
 ///
 /// Permissions are defined by [DataType], which determines the Health Kit scopes